S2 T30 version 2
Keeps the original version of this task with the bar chart, but also has a function to calculate the top 5 antidepressants prescribed in total (by quantity) and excludes Amitriptyline as asked by customer. Needs a description to be added and testing. Need to confirm that want to rank by quantity and that this is the preferred method of visualisation for the customer